# apps/derrops-cloud/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/openapi/upload-url": {
      "post": {
        "operationId": "OpenApiIndexerController_getUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a pre-signed PUT URL for uploading a spec to the OASpec bucket",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/index": {
      "post": {
        "operationId": "OpenApiIndexerController_indexFromS3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexFromS3Dto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Index a spec from S3 — runs the 6-step OASpec pipeline",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/catalogue": {
      "get": {
        "operationId": "OpenApiIndexerController_searchCatalogue",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset (default 0)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results (default 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search the platform-managed API catalogue",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/api/{apiId}/versions": {
      "get": {
        "operationId": "OpenApiIndexerController_getVersions",
        "parameters": [
          {
            "name": "apiId",
            "required": true,
            "in": "path",
            "description": "API UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated spec version history for an API",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/api/{apiId}/operations": {
      "get": {
        "operationId": "OpenApiIndexerController_getOperations",
        "parameters": [
          {
            "name": "apiId",
            "required": true,
            "in": "path",
            "description": "API UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "description": "Comma-separated HTTP method filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": false,
            "in": "query",
            "description": "Specific version or \"latest\" (default)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated operations for an API version",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/api/{apiId}/servers": {
      "get": {
        "operationId": "OpenApiIndexerController_getServers",
        "parameters": [
          {
            "name": "apiId",
            "required": true,
            "in": "path",
            "description": "API UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated servers for an API version",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/api/{apiId}/parameters": {
      "get": {
        "operationId": "OpenApiIndexerController_getParameters",
        "parameters": [
          {
            "name": "apiId",
            "required": true,
            "in": "path",
            "description": "API UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "path | query | header | cookie",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated parameters for an API version",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
    "/openapi/api/{apiId}/models": {
      "get": {
        "operationId": "OpenApiIndexerController_getModels",
        "parameters": [
          {
            "name": "apiId",
            "required": true,
            "in": "path",
            "description": "API UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "usedIn",
            "required": false,
            "in": "query",
            "description": "request | response",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paginated models for an API version",
        "tags": [
          "OpenAPI Indexer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Derrops API",
    "description": "Internal API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UploadUrlDto": {
        "type": "object",
        "properties": {}
      },
      "IndexFromS3Dto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}